Unit test work; more to come

+++ user/test/famfs_unit.cpp
@@ -5,8 +5,6 @@
 
 #include <gtest/gtest.h>
 
-#define FAMFS_UNIT_TEST
-
 extern "C" {
 #include <unistd.h>
 #include <linux/limits.h>
@@ -15,21 +13,17 @@ extern "C" {
 #include <fcntl.h>
 #include <stdlib.h>
 
-
 #include "famfs_lib.h"
+#include "famfs_lib_internal.h"
 #include "famfs_meta.h"
 #include "famfs_ioctl.h"
 #include "xrand.h"
 #include "random_buffer.h"
 }
 
-/* These are really system test calls that need an actual famfs file system mounted
- * at /mnt/famfs
- */
-
-
 /****+++++++++++++++++++++++++++++++++++++++++++++
  * NOTE THESE TESTS MUST BE RUN AS ROOT!!
+ * (perhaps we'll get around to mitigaing this...)
  */
 
 #define FAMFS_MPT "/mnt/famfs"
@@ -42,55 +36,62 @@ TEST(famfs, dummy)
 	ASSERT_EQ(0, 0);
 }
 
-
 TEST(famfs, famfs_mkfs)
 {
-	char *buf  = (char *)calloc(1, FAMFS_LOG_LEN);
 	u64 device_size = 1024 * 1024 * 1024;
 	struct famfs_superblock *sb;
 	struct famfs_log *logp;
-	mode_t mode = 0777;
-	int lfd, sfd;
-	void *addr;
 	int rc;
 
-	system("rm -rf /tmp/famfs");
+	/* Prepare a fake famfs (move changes to this block everywhere it is) */
+	if (1) {
+		char *buf  = (char *)calloc(1, FAMFS_LOG_LEN);
+		mode_t mode = 0777;
+		int lfd, sfd;
+		void *addr;
 
-	/* Create fake famfs and famfs/.meta mount point */
-	rc = mkdir("/tmp/famfs", mode);
-	ASSERT_EQ(rc, 0);
-	rc = mkdir("/tmp/famfs/.meta", mode);
-	ASSERT_EQ(rc, 0);
+		system("rm -rf /tmp/famfs");
 
-	/* Create fake log and superblock files */
-	sfd = open("/tmp/famfs/.meta/.superblock", O_RDWR | O_CREAT, 0666);
-	ASSERT_GT(sfd, 0);
-	lfd = open("/tmp/famfs/.meta/.log", O_RDWR | O_CREAT, 0666);
-	ASSERT_GT(lfd, 0);
+		/* Create fake famfs and famfs/.meta mount point */
+		rc = mkdir("/tmp/famfs", mode);
+		ASSERT_EQ(rc, 0);
+		rc = mkdir("/tmp/famfs/.meta", mode);
+		ASSERT_EQ(rc, 0);
 
-	/* Zero out the superblock and llog files */
-	rc = write(sfd, buf, FAMFS_SUPERBLOCK_SIZE);
-	ASSERT_EQ(rc, FAMFS_SUPERBLOCK_SIZE);
+		/* Create fake log and superblock files */
+		sfd = open("/tmp/famfs/.meta/.superblock", O_RDWR | O_CREAT, 0666);
+		ASSERT_GT(sfd, 0);
+		lfd = open("/tmp/famfs/.meta/.log", O_RDWR | O_CREAT, 0666);
+		ASSERT_GT(lfd, 0);
 
-	rc = write(lfd, buf, FAMFS_LOG_LEN);
-	ASSERT_EQ(rc, FAMFS_LOG_LEN);
+		/* Zero out the superblock and llog files */
+		rc = write(sfd, buf, FAMFS_SUPERBLOCK_SIZE);
+		ASSERT_EQ(rc, FAMFS_SUPERBLOCK_SIZE);
 
-	/* Mmap fake log file */
-	addr = mmap(0, FAMFS_SUPERBLOCK_SIZE, PROT_READ | PROT_WRITE, MAP_SHARED, sfd, 0);
-	ASSERT_NE(addr, MAP_FAILED);
-	logp = (struct famfs_log *)addr;
+		rc = write(lfd, buf, FAMFS_LOG_LEN);
+		ASSERT_EQ(rc, FAMFS_LOG_LEN);
+	
+		/* Mmap fake log file */
+		addr = mmap(0, FAMFS_SUPERBLOCK_SIZE, PROT_READ | PROT_WRITE, MAP_SHARED, sfd, 0);
+		ASSERT_NE(addr, MAP_FAILED);
+		sb = (struct famfs_superblock *)addr;
 
-	/* mmap fake superblock file */
-	addr = mmap(0, FAMFS_LOG_LEN, PROT_READ | PROT_WRITE, MAP_SHARED, lfd, 0);
-	ASSERT_NE(addr, MAP_FAILED);
-	sb = (struct famfs_superblock *)addr;
+		/* mmap fake superblock file */
+		addr = mmap(0, FAMFS_LOG_LEN, PROT_READ | PROT_WRITE, MAP_SHARED, lfd, 0);
+		ASSERT_NE(addr, MAP_FAILED);
+		logp = (struct famfs_log *)addr;
 
-	memset(sb, 0, FAMFS_SUPERBLOCK_SIZE);
-	memset(logp, 0, FAMFS_LOG_LEN);
+		memset(sb, 0, FAMFS_SUPERBLOCK_SIZE);
+		memset(logp, 0, FAMFS_LOG_LEN);
 
-	/* First mkfs should succeed */
-	rc = __famfs_mkfs("/dev/dax0.0", sb, logp, device_size, 0, 0);
-	ASSERT_EQ(rc, 0);
+		/* First mkfs should succeed */
+		rc = __famfs_mkfs("/dev/dax0.0", sb, logp, device_size, 0, 0);
+		ASSERT_EQ(rc, 0);
+
+		close(lfd);
+		close(sfd);
+	}
+	/****************************** end prepare fake famfs */
 
 	/* Repeat should fail because there is a valid superblock */
 	rc = __famfs_mkfs("/dev/dax0.0", sb, logp, device_size, 0, 0);
@@ -111,9 +112,6 @@ TEST(famfs, famfs_mkfs)
 	/* Repeat with force should succeed because of force */
 	rc = __famfs_mkfs("/dev/dax0.0", sb, logp, device_size, 1, 0);
 	ASSERT_EQ(rc, 0);
-
-	close(lfd);
-	close(sfd);
 
 	/* This leaves a valid superblock and log at /tmp/famfs/.meta ... */
 
@@ -190,54 +188,54 @@ TEST(famfs, famfs_open_relpath)
 	/* /tmp/famfs should already exist and have a superblock and log in it */
 	system("mkdir -p /tmp/famfs/0000/1111/2222/3333/4444/5555");
 
-	rc = __open_relpath("/tmp/bogus/path", SB_RELPATH, 1, NULL, NULL, 1);
+	rc = __open_relpath("/tmp/bogus/path", SB_RELPATH, 1, NULL, NULL, NO_LOCK, 1);
 	ASSERT_NE(rc, 0);;
 
-	rc = __open_relpath("/tmp/bogus/path", SB_RELPATH, 1, NULL, NULL, 1);
+	rc = __open_relpath("/tmp/bogus/path", SB_RELPATH, 1, NULL, NULL, NO_LOCK, 1);
 	ASSERT_NE(rc, 0);
 
 	/* Good, no ascent necessary  */
-	rc = __open_relpath("/tmp/famfs/", LOG_RELPATH, 1, NULL, NULL, 1);
+	rc = __open_relpath("/tmp/famfs/", LOG_RELPATH, 1, NULL, NULL, NO_LOCK, 1);
 	ASSERT_GT(rc, 0);
 	close(rc);
-	rc = __open_relpath("/tmp/famfs", LOG_RELPATH, 1, NULL, NULL, 1);
+	rc = __open_relpath("/tmp/famfs", LOG_RELPATH, 1, NULL, NULL, NO_LOCK, 1);
 	ASSERT_GT(rc, 0);
 	close(rc);
 
 	/* Good but deep path */
 	rc = __open_relpath("/tmp/famfs/0000/1111/2222/3333/4444/5555",
-			    LOG_RELPATH, 1, NULL, NULL, 1);
+			    LOG_RELPATH, 1, NULL, NULL, NO_LOCK, 1);
 	ASSERT_GT(rc, 0);
 	close(rc);
 
 	/* Bogus path that ascends to a real path with .meta */
 	rc = __open_relpath("/tmp/famfs/0000/1111/2222/3333/4444/5555/66666",
-			    LOG_RELPATH, 1, NULL, NULL, 1);
+			    LOG_RELPATH, 1, NULL, NULL, NO_LOCK, 1);
 	ASSERT_GT(rc, 0);
 	close(rc);
 
 	/* Deep bogus path that ascends to a real path with .meta */
 	rc = __open_relpath("/tmp/famfs/0000/1111/2222/3333/4444/5555/66666/7/6/5/4/3/2/xxx",
-			    LOG_RELPATH, 1, NULL, NULL, 1);
+			    LOG_RELPATH, 1, NULL, NULL, NO_LOCK, 1);
 	ASSERT_GT(rc, 0);
 	close(rc);
 
 	/* empty path */
-	rc = __open_relpath("", LOG_RELPATH, 1, NULL, NULL, 1);
+	rc = __open_relpath("", LOG_RELPATH, 1, NULL, NULL, NO_LOCK, 1);
 	ASSERT_LT(rc, 0);
 	close(rc);
 
 	/* "/" */
-	rc = __open_relpath("/", LOG_RELPATH, 1, NULL, NULL, 1);
+	rc = __open_relpath("/", LOG_RELPATH, 1, NULL, NULL, NO_LOCK, 1);
 	ASSERT_LT(rc, 0);
 	close(rc);
 
 	/* No "/" */
-	rc = __open_relpath("blablabla", LOG_RELPATH, 1, NULL, NULL, 1);
+	rc = __open_relpath("blablabla", LOG_RELPATH, 1, NULL, NULL, BLOCKING_LOCK, 1);
 	ASSERT_LT(rc, 0);
 	close(rc);
 	/* No "/" and spaces */
-	rc = __open_relpath("bla bla bla", LOG_RELPATH, 1, NULL, NULL, 1);
+	rc = __open_relpath("bla bla bla", LOG_RELPATH, 1, NULL, NULL, NON_BLOCKING_LOCK, 1);
 	ASSERT_LT(rc, 0);
 	close(rc);
 }
@@ -344,4 +342,93 @@ TEST(famfs, __famfs_cp)
 			0, 0, 0, 0);
 	ASSERT_EQ(rc, 1);
 
+}
+
+TEST(famfs, famfs_log)
+{
+	u64 device_size = 1024 * 1024 * 1024;
+	struct famfs_locked_log ll;
+	struct famfs_superblock *sb;
+	struct famfs_log *logp;
+	extern int mock_kmod;
+	int rc;
+	int i;
+
+	mock_kmod = 1;
+	/** can call famfs_file_alloc() and __famfs_mkdir() on our fake famfs in /tmp/famfs */
+
+	/* Prepare a fake famfs (move changes to this block everywhere it is) */
+	if (1) {
+		char *buf  = (char *)calloc(1, FAMFS_LOG_LEN);
+		mode_t mode = 0777;
+		int lfd, sfd;
+		void *addr;
+
+		system("rm -rf /tmp/famfs");
+
+		/* Create fake famfs and famfs/.meta mount point */
+		rc = mkdir("/tmp/famfs", mode);
+		ASSERT_EQ(rc, 0);
+		rc = mkdir("/tmp/famfs/.meta", mode);
+		ASSERT_EQ(rc, 0);
+
+		/* Create fake log and superblock files */
+		sfd = open("/tmp/famfs/.meta/.superblock", O_RDWR | O_CREAT, 0666);
+		ASSERT_GT(sfd, 0);
+		lfd = open("/tmp/famfs/.meta/.log", O_RDWR | O_CREAT, 0666);
+		ASSERT_GT(lfd, 0);
+
+		/* Zero out the superblock and llog files */
+		rc = write(sfd, buf, FAMFS_SUPERBLOCK_SIZE);
+		ASSERT_EQ(rc, FAMFS_SUPERBLOCK_SIZE);
+
+		rc = write(lfd, buf, FAMFS_LOG_LEN);
+		ASSERT_EQ(rc, FAMFS_LOG_LEN);
+	
+		/* Mmap fake log file */
+		addr = mmap(0, FAMFS_SUPERBLOCK_SIZE, PROT_READ | PROT_WRITE, MAP_SHARED, sfd, 0);
+		ASSERT_NE(addr, MAP_FAILED);
+		sb = (struct famfs_superblock *)addr;
+
+		/* mmap fake superblock file */
+		addr = mmap(0, FAMFS_LOG_LEN, PROT_READ | PROT_WRITE, MAP_SHARED, lfd, 0);
+		ASSERT_NE(addr, MAP_FAILED);
+		logp = (struct famfs_log *)addr;
+
+		memset(sb, 0, FAMFS_SUPERBLOCK_SIZE);
+		memset(logp, 0, FAMFS_LOG_LEN);
+
+		/* First mkfs should succeed */
+		rc = __famfs_mkfs("/dev/dax0.0", sb, logp, device_size, 0, 0);
+		ASSERT_EQ(rc, 0);
+
+		close(lfd);
+		close(sfd);
+	}
+
+	rc = famfs_init_locked_log(&ll, "/tmp/famfs", 1);
+	ASSERT_EQ(rc, 0);
+
+	for (i = 0; i < 10; i++) {
+		char filename[64];
+		int fd;
+		sprintf(filename, "/tmp/famfs/%04d", i);
+		fd = __famfs_mkfile(&ll, filename, 0, 0, 0, 1048576, 0);
+		ASSERT_GT(fd, 0);
+		close(fd);
+	}
+
+	/* TODO: nested dirs and files to fill up the log */
+	for (i = 0; i < 100; i++) {
+		char dirname[64];
+		sprintf(dirname, "/tmp/famfs/dir%04d", i);
+		rc = __famfs_mkdir(&ll, dirname, 0, 0, 0, 0);
+		ASSERT_EQ(rc, 0);
+	}
+	rc = __famfs_logplay(logp, "/tmp/famfs", 0, 0, 3);
+	ASSERT_EQ(rc, 0);
+	//famfs_print_log_stats("famfs_log test", )
+
+	rc = famfs_fsck_scan(sb, logp, 1, 3);
+	ASSERT_EQ(rc, 0);
 }
